Dot(.) splitting should be optional and controlable with params
In my project, I needed a feature to avoid splitting on dot(.). Because due to this my translations were malfunctioning as if I've translation as "I am ABC test. And I want to go XYZ place.". The existing codebase was splitting it on separate translations for "I am ABC test" and "I want to go XYZ". Whereas there was no translation for them in my system, there was a translation for a complete sentence. 

Now if someone also comes in similar circumstances, they can call it like this `translate(key, {}, false)`. And due to this 3rd parameter, the translate function would not consider dot spliting.